+++ mm/debug.c
@@ -96,7 +96,7 @@ static void __dump_folio(const struct folio *folio, const struct page *page,
 
 #ifdef CONFIG_MEMCG
 	if (folio->memcg_data)
-		pr_warn("memcg:%lx\n", folio->memcg_data);
+		pr_warn("memcg:%lx\n", (unsigned long)folio->memcg_data);
 #endif
 	if (folio_test_ksm(folio))
 		type = "ksm ";
